+++ firmware-arduino/src/main.cpp
@@ -190,10 +190,10 @@ void getAuthTokenFromNVS() {
 }
 
 void setupWiFi() {
-  WifiManager.startBackgroundTask(
-      "ELATO-DEVICE"); // Run the background task to take care of our Wifi
   WifiManager.fallbackToSoftAp(
       true); // Run a SoftAP if no known AP can be reached
+  WifiManager.startBackgroundTask(
+      "VOICE-DEVICE"); // Run the background task to take care of our Wifi
   WifiManager.attachWebServer(&webServer); // Attach our API to the Webserver
   WifiManager.attachUI();                  // Attach the UI to the Webserver